# apps/web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivationView": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "planning_revision": {
            "title": "Planning Revision",
            "type": "integer"
          },
          "publication_state": {
            "title": "Publication State",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "publication_state",
          "planning_revision"
        ],
        "title": "ActivationView",
        "type": "object"
      },
...
      "Candidate": {
        "additionalProperties": false,
        "properties": {
          "blocks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Block"
            },
            "title": "Blocks",
            "type": "array"
          },
          "constraint_report": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Violation"
            },
            "title": "Constraint Report",
            "type": "array"
          },
          "planning_revision": {
            "minimum": 0,
            "title": "Planning Revision",
            "type": "integer"
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Score"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot_hash": {
            "title": "Snapshot Hash",
            "type": "string"
          },
          "solver_metadata": {
            "$ref": "#/components/schemas/SolverMetadata"
          },
          "source_policy": {
            "title": "Source Policy",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPTIMAL",
              "FEASIBLE",
              "INFEASIBLE",
              "UNKNOWN",
              "MODEL_INVALID"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "snapshot_hash",
          "planning_revision",
          "status",
          "source_policy"
        ],
        "title": "Candidate",
        "type": "object"
      },
...
      "DateDeadlineChange": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "new_bound": {
            "title": "New Bound",
            "type": "string"
          },
          "old_bound": {
            "title": "Old Bound",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "date",
          "old_bound",
          "new_bound"
        ],
        "title": "DateDeadlineChange",
        "type": "object"
      },
...
      "GenerateResponse": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "planning_revision": {
            "title": "Planning Revision",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "state",
          "planning_revision"
        ],
        "title": "GenerateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobView": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Candidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "planning_revision": {
            "title": "Planning Revision",
            "type": "integer"
          },
          "proposal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposal Id"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "planning_revision"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "OriginalTimeInput": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "fold": {
            "anyOf": [
              {
                "enum": [
                  0,
                  1
                ],
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fold"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "field",
          "value",
          "timezone"
        ],
        "title": "OriginalTimeInput",
        "type": "object"
      },
...
      "ProposalView": {
        "additionalProperties": false,
        "properties": {
          "activated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activated At"
          },
          "candidate": {
            "$ref": "#/components/schemas/Candidate"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "planning_revision": {
            "title": "Planning Revision",
            "type": "integer"
          },
          "publication_state": {
            "title": "Publication State",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "planning_revision",
          "candidate",
          "publication_state"
        ],
        "title": "ProposalView",
        "type": "object"
      },
...
      "RevisionCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "RevisionCommand",
        "type": "object"
      },
...
      "RoundingLoss": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "original": {
            "format": "date-time",
            "title": "Original",
            "type": "string"
          },
          "rounded": {
            "format": "date-time",
            "title": "Rounded",
            "type": "string"
          },
          "seconds": {
            "minimum": 0,
            "title": "Seconds",
            "type": "number"
          }
        },
        "required": [
          "field",
          "original",
          "rounded",
          "seconds"
        ],
        "title": "RoundingLoss",
        "type": "object"
      },
...
      "TimeInputView": {
        "additionalProperties": false,
        "properties": {
          "original_time_inputs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OriginalTimeInput"
            },
            "title": "Original Time Inputs",
            "type": "array"
          },
          "rounding_losses": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RoundingLoss"
            },
            "title": "Rounding Losses",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "timezone"
        ],
        "title": "TimeInputView",
        "type": "object"
      },
      "TimezoneCommand": {
        "additionalProperties": false,
        "properties": {
          "confirmed": {
            "const": true,
            "title": "Confirmed",
            "type": "boolean"
          },
          "expected_revision": {
            "minimum": 0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "preview_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Preview Hash",
            "type": "string"
          },
          "timezone": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "timezone",
          "preview_hash",
          "confirmed"
        ],
        "title": "TimezoneCommand",
        "type": "object"
      },
      "TimezonePreview": {
        "additionalProperties": false,
        "properties": {
          "date_deadlines": {
            "items": {
              "$ref": "#/components/schemas/DateDeadlineChange"
            },
            "title": "Date Deadlines",
            "type": "array"
          },
          "old_timezone": {
            "title": "Old Timezone",
            "type": "string"
          },
          "preserved_available_windows": {
            "title": "Preserved Available Windows",
            "type": "integer"
          },
          "preserved_fixed_events": {
            "title": "Preserved Fixed Events",
            "type": "integer"
          },
          "preview_hash": {
            "title": "Preview Hash",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "weekly_rules_follow_new_timezone": {
            "title": "Weekly Rules Follow New Timezone",
            "type": "boolean"
          }
        },
        "required": [
          "old_timezone",
          "timezone",
          "revision",
          "date_deadlines",
          "preserved_fixed_events",
          "preserved_available_windows",
          "weekly_rules_follow_new_timezone",
          "preview_hash"
        ],
        "title": "TimezonePreview",
        "type": "object"
      },
      "TimezoneSaved": {
        "additionalProperties": false,
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "revision",
          "timezone"
        ],
        "title": "TimezoneSaved",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Violation": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "facts": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Facts",
            "type": "object"
          },
          "related_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Related Ids",
            "type": "array"
          }
        },
        "required": [
          "code"
        ],
        "title": "Violation",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Adaptive Planner",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/time-inputs": {
      "get": {
        "operationId": "job_time_inputs_api_v1_jobs__job_id__time_inputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeInputView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Time Inputs",
        "tags": [
          "plans"
        ]
      }
    },
...
    "/api/v1/proposals/{proposal_id}": {
      "get": {
        "operationId": "get_proposal_api_v1_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Proposal",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/proposals/{proposal_id}/activate": {
      "post": {
        "operationId": "activate_api_v1_proposals__proposal_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate",
        "tags": [
          "plans"
        ]
      }
    },
...
    "/api/v1/proposals/{proposal_id}/time-inputs": {
      "get": {
        "operationId": "proposal_time_inputs_api_v1_proposals__proposal_id__time_inputs_get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeInputView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Proposal Time Inputs",
        "tags": [
          "plans"
        ]
      }
    },
...
    "/api/v1/replans": {
      "post": {
        "operationId": "generate_api_v1_replans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/v1/settings/timezone": {
      "put": {
        "operationId": "change_api_v1_settings_timezone_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimezoneCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezoneSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/timezone-preview": {
      "get": {
        "operationId": "preview_api_v1_settings_timezone_preview_get",
        "parameters": [
          {
            "in": "query",
            "name": "timezone",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Timezone",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimezonePreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}